+++ engine/src/render/passes/lens_flare_pass.zig
@@ -62,7 +62,7 @@ pub fn runPipeline(
             @ptrCast(&self.lens_flare_job_contexts[stripe_index]),
             &parent_job,
         );
-        if (!self.job_system.?.submitJobAuto(&self.color_grade_jobs[stripe_index])) {
+        if (!self.job_system.?.submitJobWithClass(&self.color_grade_jobs[stripe_index], .background)) {
             CtxType.run(@ptrCast(&self.lens_flare_job_contexts[stripe_index]));
         }
     }